+++ src/lib/components/chat/components/renderer/no-paragraph.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  export let raw: string;
+</script>
+
+{raw}
